Detalle y Eliminar agregados correctamente

+++ src/main/java/com/example/demo/service/ProductoService.java
@@ -6,7 +6,7 @@ import com.example.demo.entity.ProductoEntity;
 
 public interface ProductoService {
 	List<ProductoEntity> buscarTodosProductos();
-	ProductoEntity buscarProductoPorId(Integer id);
+	ProductoEntity buscarProductoPorId(Long id);
 	void save(ProductoEntity producto); 
 
 }
